Competition Data
Interactions_and_impressions.csv : Contains the training set, describing implicit preferences expressed by the users.
user_id : identifier of the user
item_id : identifier of the item (TV series)
impression_list : string containing the items that were present on the screen when the user interacted with the item in column item_id. Not all interactions have a corresponding impressions list.
data : "0" if the user watched the item, "1" if the user opened the item details page.
Note that there are multiple interactions between the same user and item when a user watches multiple episodes of a TV series (if a user has watched 5 episodes there will be 5 interactions with that item_id).